pulumi/lib: shave some sharp edges off the AWS Lambda API.
