+++ ts/pulumi/zemn.me/hello_world/hello_world.ts
@@ -1,35 +1,13 @@
 import * as aws from "@pulumi/aws";
 import { ComponentResource, ComponentResourceOptions } from "@pulumi/pulumi";
 
+import { LambdaFunction } from "#root/ts/pulumi/lib/lambda_function.js";
 import { mergeTags, TagSet, tagTrue } from "#root/ts/pulumi/lib/tags.js";
 import { HelloWorldImage } from "#root/ts/pulumi/zemn.me/hello_world/HelloWorldImage.js";
 
 export interface Args {
     tags?: TagSet;
 }
-
-function clampLambdaFunctionName(name: string): string {
-  // Define constraints
-  const maxLength = 64;
-  const allowedCharactersRegex = /^[a-zA-Z0-9_-]+$/;
-
-  // Trim the string to the maximum allowed length
-  let clampedName = name.slice(0, maxLength);
-
-  // Remove invalid characters
-  clampedName = clampedName
-    .split('')
-    .filter(char => allowedCharactersRegex.test(char))
-    .join('');
-
-  // Ensure the name is at least one character long
-  if (clampedName.length === 0) {
-    throw new Error("Lambda function name must contain at least one valid character.");
-  }
-
-  return clampedName;
-}
-
 
 export class LambdaHelloWorld extends ComponentResource {
     constructor(
@@ -63,8 +41,7 @@ export class LambdaHelloWorld extends ComponentResource {
 				Service: "lambda.amazonaws.com",
 			}),
 		});
-        // Create the Lambda function using the container image
-        const lambda = new aws.lambda.Function(clampLambdaFunctionName(`${name}_lambda`), {
+        const lambda = new LambdaFunction(`${name}_lambda`, {
             packageType: "Image",
 			// e.g. --
 			// https://658613637108.dkr.ecr.us-east-1.amazonaws.com/repo-fe6824f@sha256:c61c4b619733eb59d7c68a34d7b56e562894c9d777f26e982bc453bbcf57ab5b
@@ -83,7 +60,7 @@ export class LambdaHelloWorld extends ComponentResource {
 
         // Export the Lambda function name
         this.registerOutputs({
-            lambdaFunctionName: lambda.name,
+            lambdaFunctionName: lambda.function.name,
         });
     }
 }
